refactoring/clean up
