+++ packages/jasmine/models/error.messages.model.ts
@@ -1,5 +1,5 @@
 import { ExpectedResult, ActualResult } from './matcher.model';
-import { MatchersTypes } from './matcher';
+import { MatchersTypes } from '../matcher';
 
 export type ErrorMessageCallback = (
     expectedResult: ExpectedResult,
