+++ packages/jasmine/store.ts
@@ -1,13 +1,4 @@
-import { Describers, NextDescriberArguments } from './describe.model';
-
-export interface Store {
-    activeDescriberId: string;
-    isDescriberFormingInProgress: boolean;
-    describers: Describers;
-    rootDescribersId: Array<string>;
-    activeTestCaseIndex: number;
-    nextDescriberArguments: Array<NextDescriberArguments>;
-}
+import { Store } from './models/store.model';
 
 export const store: Store = {
     activeDescriberId: null,
